Use the jquery file upload widget when editing people
Selecting an avatar now triggers the file upload to happen immediately. It
actually triggers the whole form to be POSTed, so changes to the name will
also be persisted. It looks as though the avatar isn't uploaded a second time
when you hit the "Update person" button, but any change to the name is sent to
the server.

I've only enabled the jquery file upload when editing a person, as making it work when creating new people is more complicated.